+++ coupleDayProject/Common/Constants.swift
@@ -34,96 +34,6 @@ struct CommonSize {
 	? 120.0 : 100.0
 }
 
-struct RealmManager {
-	//         try! FileManager.default.removeItem(at:Realm.Configuration.defaultConfiguration.fileURL!)
-	
-	static let shared: RealmManager = .init()
-	
-	private var realm: Realm {
-		let container = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: "group.ungchun.coupleDayProject")
-		let realmURL = container?.appendingPathComponent("default.realm")
-		let config = Realm.Configuration(fileURL: realmURL, schemaVersion: 1)
-		do {
-			return try Realm(configuration: config)
-		} catch let error as NSError {
-			print(error.debugDescription)
-			fatalError("Can't continue further, no Realm available")
-		}
-	}
-	
-	func getUserDatas() -> [RealmUserModel] {
-		Array(realm.objects(RealmUserModel.self))
-	}
-	
-	func getImageDatas() -> [RealmImageModel] {
-		Array(realm.objects(RealmImageModel.self))
-	}
-	
-	func writeUserData(userData: RealmUserModel) {
-		try? realm.write({
-			realm.add(userData)
-		})
-	}
-	
-	func writeImageData(imageData: RealmImageModel) {
-		try? realm.write({
-			realm.add(imageData)
-		})
-	}
-	
-	func updateBeginCoupleDay(datePicker: UIDatePicker) {
-		try? realm.write({
-			if RealmManager.shared.getUserDatas().first!.zeroDayStartCheck {
-				RealmManager.shared.getUserDatas().first!.beginCoupleDay = Int(datePicker.date.toString.toDate.millisecondsSince1970)
-			} else {
-				RealmManager.shared.getUserDatas().first!.beginCoupleDay = Int(
-					Calendar.current.date(
-						byAdding: .day,
-						value: -1,
-						to: datePicker.date.toString.toDate
-					)!.millisecondsSince1970
-				)
-			}
-		})
-		WidgetCenter.shared.reloadAllTimelines()
-	}
-	
-	func updateBirthDay(datePicker: UIDatePicker) {
-		initBirthDayAnniversaryModel(
-			dateValue: Int(datePicker.date.toString.toDate.millisecondsSince1970)
-		)
-		try? realm.write({
-			RealmManager.shared.getUserDatas().first!.birthDay = Int(datePicker.date.toString.toDate.millisecondsSince1970)
-		})
-	}
-	
-	func updateMainImage(mainImage: UIImage) {
-		try? realm.write({
-			RealmManager.shared.getImageDatas().first!.homeMainImage =
-			(mainImage.pngData()?.count)! > 10000000
-			? resizeImage(image: mainImage, newWidth: 1000).pngData()
-			: mainImage.jpegData(compressionQuality: 0.5)
-		})
-		WidgetCenter.shared.reloadAllTimelines()
-	}
-	
-	func updateMyProfileImage(myProfileImage: UIImage) {
-		try? realm.write({
-			RealmManager.shared.getImageDatas().first!.myProfileImage =
-			(myProfileImage.pngData()?.count)! > 10000000
-			? resizeImage(image: myProfileImage, newWidth: 1000).pngData()
-			: myProfileImage.jpegData(compressionQuality: 0.5)
-		})
-	}
-	
-	func updatePartnerProfileImage(partnerProfileImage: UIImage) {
-		try? realm.write({
-			RealmManager.shared.getImageDatas().first!.partnerProfileImage = (partnerProfileImage.pngData()?.count)! > 10000000
-			? resizeImage(image: partnerProfileImage, newWidth: 1000).pngData() : partnerProfileImage.jpegData(compressionQuality: 0.5)
-		})
-	}
-}
-
 // UIImage 객체의 사진용량 줄이기
 
 func resizeImage(image: UIImage, newWidth: CGFloat) -> UIImage {
