+++ coupleDayProject/Presentation/ViewModel/CoupleTabViewModel.swift
@@ -33,7 +33,7 @@ final class CoupleTabViewModel {
 	func updateBeginCoupleDay() {
 		let nowDayDataString = Date().toString
 		let nowDayDataDate: Date = nowDayDataString.toDate
-		let minus = Int(nowDayDataDate.millisecondsSince1970)-RealmManager.shared.getUserDatas().first!.beginCoupleDay
+		let minus = Int(nowDayDataDate.millisecondsSince1970)-RealmService.shared.getUserDatas().first!.beginCoupleDay
 		self.beginCoupleDay.value = String(describing: minus / 86400000)
 		
 		NotificationCenter.default.post(
@@ -44,33 +44,33 @@ final class CoupleTabViewModel {
 	}
 	
 	func updateHomeMainImage() {
-		guard let homeMainImage = RealmManager.shared.getImageDatas().first!.homeMainImage else { return }
+		guard let homeMainImage = RealmService.shared.getImageDatas().first!.homeMainImage else { return }
 		self.homeMainImageData.value = homeMainImage
 	}
 	
 	func updateMyProfileIcon() {
 		let isDark = UserDefaults.standard.bool(forKey: "darkModeState")
-		if RealmManager.shared.getImageDatas().first!.myProfileImage == nil {
+		if RealmService.shared.getImageDatas().first!.myProfileImage == nil {
 			if isDark {
 				self.myProfileImageData.value = (UIImage(named: "smile_dark")?.jpegData(compressionQuality: 0.5))!
 			} else {
 				self.myProfileImageData.value = (UIImage(named: "smile_white")?.jpegData(compressionQuality: 0.5))!
 			}
 		} else {
-			self.myProfileImageData.value = RealmManager.shared.getImageDatas().first!.myProfileImage!
+			self.myProfileImageData.value = RealmService.shared.getImageDatas().first!.myProfileImage!
 		}
 	}
 	
 	func updatePartnerProfileIcon() {
 		let isDark = UserDefaults.standard.bool(forKey: "darkModeState")
-		if RealmManager.shared.getImageDatas().first!.partnerProfileImage == nil {
+		if RealmService.shared.getImageDatas().first!.partnerProfileImage == nil {
 			if isDark {
 				self.partnerProfileImageData.value = (UIImage(named: "smile_dark")?.jpegData(compressionQuality: 0.5))!
 			} else {
 				self.partnerProfileImageData.value = (UIImage(named: "smile_white")?.jpegData(compressionQuality: 0.5))!
 			}
 		} else {
-			self.partnerProfileImageData.value = RealmManager.shared.getImageDatas().first!.partnerProfileImage!
+			self.partnerProfileImageData.value = RealmService.shared.getImageDatas().first!.partnerProfileImage!
 		}
 	}
 }
